Refactoring 1: Cleanup profiles and move conversion helpers methods to utils (#248)
* Refactorings except common.gp

* Just common.go changes
