+++ cmd/data.go
@@ -8,8 +8,10 @@ import (
 	"path"
 	"time"
 
+	"github.com/cloudspannerecosystem/harbourbridge/common/utils"
 	"github.com/cloudspannerecosystem/harbourbridge/conversion"
 	"github.com/cloudspannerecosystem/harbourbridge/internal"
+	"github.com/cloudspannerecosystem/harbourbridge/profiles"
 	"github.com/google/subcommands"
 )
 
@@ -64,7 +66,7 @@ func (cmd *DataCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface
 		}
 	}()
 
-	sourceProfile, err := NewSourceProfile(cmd.sourceProfile, cmd.source)
+	sourceProfile, err := profiles.NewSourceProfile(cmd.sourceProfile, cmd.source)
 	if err != nil {
 		return subcommands.ExitUsageError
 	}
@@ -73,23 +75,23 @@ func (cmd *DataCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface
 		return subcommands.ExitUsageError
 	}
 
-	targetProfile, err := NewTargetProfile(cmd.targetProfile)
+	targetProfile, err := profiles.NewTargetProfile(cmd.targetProfile)
 	if err != nil {
 		return subcommands.ExitUsageError
 	}
 	targetDb := targetProfile.ToLegacyTargetDb()
 
 	dumpFilePath := ""
-	if sourceProfile.ty == SourceProfileTypeFile && (sourceProfile.file.format == "" || sourceProfile.file.format == "dump") {
-		dumpFilePath = sourceProfile.file.path
+	if sourceProfile.Ty == profiles.SourceProfileTypeFile && (sourceProfile.File.Format == "" || sourceProfile.File.Format == "dump") {
+		dumpFilePath = sourceProfile.File.Path
 	}
-	ioHelper := conversion.NewIOStreams(driverName, dumpFilePath)
+	ioHelper := utils.NewIOStreams(driverName, dumpFilePath)
 	if ioHelper.SeekableIn != nil {
 		defer ioHelper.In.Close()
 	}
 
 	now := time.Now()
-	project, instance, dbName, err := getResourceIds(ctx, targetProfile, now, driverName, ioHelper.Out)
+	project, instance, dbName, err := profiles.GetResourceIds(ctx, targetProfile, now, driverName, ioHelper.Out)
 	if err != nil {
 		return subcommands.ExitUsageError
 	}
@@ -110,13 +112,13 @@ func (cmd *DataCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface
 		return subcommands.ExitUsageError
 	}
 
-	adminClient, err := conversion.NewDatabaseAdminClient(ctx)
+	adminClient, err := utils.NewDatabaseAdminClient(ctx)
 	if err != nil {
-		err = fmt.Errorf("can't create admin client: %w", conversion.AnalyzeError(err, dbURI))
+		err = fmt.Errorf("can't create admin client: %w", utils.AnalyzeError(err, dbURI))
 		return subcommands.ExitFailure
 	}
 	defer adminClient.Close()
-	client, err := conversion.GetClient(ctx, dbURI)
+	client, err := utils.GetClient(ctx, dbURI)
 	if err != nil {
 		err = fmt.Errorf("can't create client for db %s: %v", dbURI, err)
 		return subcommands.ExitFailure
@@ -129,7 +131,7 @@ func (cmd *DataCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface
 		return subcommands.ExitFailure
 	}
 
-	bw, err := conversion.DataConv(driverName, getSQLConnectionStr(sourceProfile), &ioHelper, client, conv, true, getSchemaSampleSize(sourceProfile))
+	bw, err := conversion.DataConv(driverName, profiles.GetSQLConnectionStr(sourceProfile), &ioHelper, client, conv, true, profiles.GetSchemaSampleSize(sourceProfile))
 	if err != nil {
 		err = fmt.Errorf("can't finish data conversion for db %s: %v", dbURI, err)
 		return subcommands.ExitFailure
@@ -140,7 +142,7 @@ func (cmd *DataCmd) Execute(ctx context.Context, f *flag.FlagSet, _ ...interface
 			return subcommands.ExitFailure
 		}
 	}
-	banner := conversion.GetBanner(now, dbURI)
+	banner := utils.GetBanner(now, dbURI)
 	conversion.Report(driverName, bw.DroppedRowsByTable(), ioHelper.BytesRead, banner, conv, cmd.filePrefix+reportFile, ioHelper.Out)
 	conversion.WriteBadData(bw, conv, banner, cmd.filePrefix+badDataFile, ioHelper.Out)
 	return subcommands.ExitSuccess
