+++ conversion/conversion.go
@@ -25,16 +25,12 @@ package conversion
 import (
 	"bufio"
 	"context"
-	"crypto/rand"
 	"database/sql"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"log"
-	"net/url"
 	"os"
-	"os/exec"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -43,20 +39,15 @@ import (
 
 	sp "cloud.google.com/go/spanner"
 	database "cloud.google.com/go/spanner/admin/database/apiv1"
-	instance "cloud.google.com/go/spanner/admin/instance/apiv1"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	dydb "github.com/aws/aws-sdk-go/service/dynamodb"
-	"golang.org/x/crypto/ssh/terminal"
-	"google.golang.org/api/iterator"
-	"google.golang.org/api/option"
 	adminpb "google.golang.org/genproto/googleapis/spanner/admin/database/v1"
-	instancepb "google.golang.org/genproto/googleapis/spanner/admin/instance/v1"
-
-	"cloud.google.com/go/storage"
 
 	"github.com/cloudspannerecosystem/harbourbridge/common/constants"
+	"github.com/cloudspannerecosystem/harbourbridge/common/utils"
 	"github.com/cloudspannerecosystem/harbourbridge/internal"
+	"github.com/cloudspannerecosystem/harbourbridge/profiles"
 	"github.com/cloudspannerecosystem/harbourbridge/sources/common"
 	"github.com/cloudspannerecosystem/harbourbridge/sources/dynamodb"
 	"github.com/cloudspannerecosystem/harbourbridge/sources/mysql"
@@ -82,7 +73,7 @@ var (
 //  - Driver is DynamoDB or a dump file mode.
 //  - This function is called as part of the legacy global CLI flag mode. (This string is constructed from env variables later on)
 // When using source-profile, the sqlConnectionStr is constructed from the input params.
-func SchemaConv(driver, sqlConnectionStr, targetDb string, ioHelper *IOStreams, schemaSampleSize int64) (*internal.Conv, error) {
+func SchemaConv(driver, sqlConnectionStr, targetDb string, ioHelper *utils.IOStreams, schemaSampleSize int64) (*internal.Conv, error) {
 	switch driver {
 	case constants.POSTGRES, constants.MYSQL, constants.DYNAMODB:
 		return schemaFromDatabase(driver, sqlConnectionStr, targetDb, schemaSampleSize)
@@ -99,7 +90,7 @@ func SchemaConv(driver, sqlConnectionStr, targetDb string, ioHelper *IOStreams, 
 //  - Driver is DynamoDB or a dump file mode.
 //  - This function is called as part of the legacy global CLI flag mode. (This string is constructed from env variables later on)
 // When using source-profile, the sqlConnectionStr and schemaSampleSize are constructed from the input params.
-func DataConv(driver, sqlConnectionStr string, ioHelper *IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool, schemaSampleSize int64) (*spanner.BatchWriter, error) {
+func DataConv(driver, sqlConnectionStr string, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool, schemaSampleSize int64) (*spanner.BatchWriter, error) {
 	config := spanner.BatchWriterConfig{
 		BytesLimit: 100 * 1000 * 1000,
 		WriteLimit: 40,
@@ -125,14 +116,14 @@ func connectionConfig(driver string, sqlConnectionStr string) (interface{}, erro
 		// If empty, this is called as part of the legacy mode witih global CLI flags.
 		// When using source-profile mode is used, the sqlConnectionStr is already populated.
 		if sqlConnectionStr == "" {
-			return generatePGSQLConnectionStr()
+			return profiles.GeneratePGSQLConnectionStr()
 		}
 		return sqlConnectionStr, nil
 	case constants.MYSQL:
 		// If empty, this is called as part of the legacy mode witih global CLI flags.
 		// When using source-profile mode is used, the sqlConnectionStr is already populated.
 		if sqlConnectionStr == "" {
-			return generateMYSQLConnectionStr()
+			return profiles.GenerateMYSQLConnectionStr()
 		}
 		return sqlConnectionStr, nil
 	case constants.DYNAMODB:
@@ -140,46 +131,6 @@ func connectionConfig(driver string, sqlConnectionStr string) (interface{}, erro
 	default:
 		return "", fmt.Errorf("driver %s not supported", driver)
 	}
-}
-
-func generatePGSQLConnectionStr() (string, error) {
-	server := os.Getenv("PGHOST")
-	port := os.Getenv("PGPORT")
-	user := os.Getenv("PGUSER")
-	dbname := os.Getenv("PGDATABASE")
-	if server == "" || port == "" || user == "" || dbname == "" {
-		fmt.Printf("Please specify host, port, user and database using PGHOST, PGPORT, PGUSER and PGDATABASE environment variables\n")
-		return "", fmt.Errorf("could not connect to source database")
-	}
-	password := os.Getenv("PGPASSWORD")
-	if password == "" {
-		password = GetPassword()
-	}
-	return GetPGSQLConnectionStr(server, port, user, password, dbname), nil
-}
-
-func GetPGSQLConnectionStr(server, port, user, password, dbname string) string {
-	return fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=disable", server, port, user, password, dbname)
-}
-
-func generateMYSQLConnectionStr() (string, error) {
-	server := os.Getenv("MYSQLHOST")
-	port := os.Getenv("MYSQLPORT")
-	user := os.Getenv("MYSQLUSER")
-	dbname := os.Getenv("MYSQLDATABASE")
-	if server == "" || port == "" || user == "" || dbname == "" {
-		fmt.Printf("Please specify host, port, user and database using MYSQLHOST, MYSQLPORT, MYSQLUSER and MYSQLDATABASE environment variables\n")
-		return "", fmt.Errorf("could not connect to source database")
-	}
-	password := os.Getenv("MYSQLPWD")
-	if password == "" {
-		password = GetPassword()
-	}
-	return GetMYSQLConnectionStr(server, port, user, password, dbname), nil
-}
-
-func GetMYSQLConnectionStr(server, port, user, password, dbname string) string {
-	return fmt.Sprintf("%s:%s@tcp(%s:%s)/%s", user, password, server, port, dbname)
 }
 
 func getDbNameFromSQLConnectionStr(driver, sqlConnectionStr string) string {
@@ -244,100 +195,10 @@ func getDynamoDBClientConfig() (*aws.Config, error) {
 	return &cfg, nil
 }
 
-// IOStreams is a struct that contains the file descriptor for dumpFile.
-type IOStreams struct {
-	In, SeekableIn, Out *os.File
-	BytesRead           int64
-}
-
-// downloadFromGCS returns the dump file that is downloaded from GCS
-func downloadFromGCS(bucketName string, filePath string) (*os.File, error) {
-	ctx := context.Background()
-
-	client, err := storage.NewClient(ctx)
-	if err != nil {
-		fmt.Printf("Failed to create GCS client for bucket %q", bucketName)
-		log.Fatal(err)
-	}
-	defer client.Close()
-
-	bucket := client.Bucket(bucketName)
-	rc, err := bucket.Object(filePath).NewReader(ctx)
-	if err != nil {
-		fmt.Printf("readFile: unable to open file from bucket %q, file %q: %v", bucketName, filePath, err)
-		log.Fatal(err)
-		return nil, err
-	}
-	defer rc.Close()
-	r := bufio.NewReader(rc)
-
-	tmpfile, err := ioutil.TempFile("", "harbourbridge.gcs.data")
-	if err != nil {
-		fmt.Printf("saveFile: unable to open temporary file to save dump file from GCS bucket %v", err)
-		log.Fatal(err)
-		return nil, err
-	}
-	syscall.Unlink(tmpfile.Name()) // File will be deleted when this process exits.
-
-	fmt.Printf("\nDownloading dump file from GCS bucket %s, path %s\n", bucketName, filePath)
-	buffer := make([]byte, 1024)
-	for {
-		// read a chunk
-		n, err := r.Read(buffer[:cap(buffer)])
-
-		if err != nil && err != io.EOF {
-			fmt.Printf("readFile: unable to read entire dump file from bucket %s, file %s: %v", bucketName, filePath, err)
-			log.Fatal(err)
-			return nil, err
-		}
-		if n == 0 && err == io.EOF {
-			break
-		}
-
-		// write a chunk
-		if _, err = tmpfile.Write(buffer[:n]); err != nil {
-			fmt.Printf("saveFile: unable to save read data from bucket %s, file %s: %v", bucketName, filePath, err)
-			log.Fatal(err)
-		}
-	}
-
-	return tmpfile, nil
-}
-
-// NewIOStreams returns a new IOStreams struct such that input stream is set
-// to open file descriptor for dumpFile if driver is PGDUMP or MYSQLDUMP.
-// Input stream defaults to stdin. Output stream is always set to stdout.
-func NewIOStreams(driver string, dumpFile string) IOStreams {
-	io := IOStreams{In: os.Stdin, Out: os.Stdout}
-	u, err := url.Parse(dumpFile)
-	if err != nil {
-		fmt.Printf("parseFilePath: unable parse file path for dumpfile %s", dumpFile)
-		log.Fatal(err)
-	}
-	if (driver == constants.PGDUMP || driver == constants.MYSQLDUMP) && dumpFile != "" {
-		fmt.Printf("\nLoading dump file from path: %s\n", dumpFile)
-		var f *os.File
-		var err error
-		if u.Scheme == "gs" {
-			bucketName := u.Host
-			filePath := u.Path[1:] // removes "/" from beginning of path
-			f, err = downloadFromGCS(bucketName, filePath)
-		} else {
-			f, err = os.Open(dumpFile)
-		}
-		if err != nil {
-			fmt.Printf("\nError reading dump file: %v err:%v\n", dumpFile, err)
-			log.Fatal(err)
-		}
-		io.In = f
-	}
-	return io
-}
-
-func schemaFromDump(driver string, targetDb string, ioHelper *IOStreams) (*internal.Conv, error) {
+func schemaFromDump(driver string, targetDb string, ioHelper *utils.IOStreams) (*internal.Conv, error) {
 	f, n, err := getSeekable(ioHelper.In)
 	if err != nil {
-		printSeekError(driver, err, ioHelper.Out)
+		utils.PrintSeekError(driver, err, ioHelper.Out)
 		return nil, fmt.Errorf("can't get seekable input file")
 	}
 	ioHelper.SeekableIn = f
@@ -357,7 +218,7 @@ func schemaFromDump(driver string, targetDb string, ioHelper *IOStreams) (*inter
 	return conv, nil
 }
 
-func dataFromDump(driver string, config spanner.BatchWriterConfig, ioHelper *IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool) (*spanner.BatchWriter, error) {
+func dataFromDump(driver string, config spanner.BatchWriterConfig, ioHelper *utils.IOStreams, client *sp.Client, conv *internal.Conv, dataOnly bool) (*spanner.BatchWriter, error) {
 	// TODO: refactor of the way we handle getSeekable
 	// to avoid the code duplication here
 	if !dataOnly {
@@ -371,7 +232,7 @@ func dataFromDump(driver string, config spanner.BatchWriterConfig, ioHelper *IOS
 		// changes in showing progress for data migration.
 		f, n, err := getSeekable(ioHelper.In)
 		if err != nil {
-			printSeekError(driver, err, ioHelper.Out)
+			utils.PrintSeekError(driver, err, ioHelper.Out)
 			return nil, fmt.Errorf("can't get seekable input file")
 		}
 		ioHelper.SeekableIn = f
@@ -445,7 +306,7 @@ func Report(driver string, badWrites map[string]int64, BytesRead int64, banner s
 func getSeekable(f *os.File) (*os.File, int64, error) {
 	_, err := f.Seek(0, 0)
 	if err == nil { // Stdin is seekable, let's just use that. This happens when you run 'cmd < file'.
-		n, err := getSize(f)
+		n, err := utils.GetFileSize(f)
 		return f, n, err
 	}
 	internal.VerbosePrintln("Creating a tmp file with a copy of stdin because stdin is not seekable.")
@@ -468,7 +329,7 @@ func getSeekable(f *os.File) (*os.File, int64, error) {
 	if err != nil {
 		return nil, 0, fmt.Errorf("can't reset file offset: %w", err)
 	}
-	n, _ := getSize(fcopy)
+	n, _ := utils.GetFileSize(fcopy)
 	return fcopy, n, nil
 }
 
@@ -488,7 +349,7 @@ func VerifyDb(ctx context.Context, adminClient *database.DatabaseAdminClient, db
 func CheckExistingDb(ctx context.Context, adminClient *database.DatabaseAdminClient, dbURI string) (bool, error) {
 	_, err := adminClient.GetDatabase(ctx, &adminpb.GetDatabaseRequest{Name: dbURI})
 	if err != nil {
-		if containsAny(strings.ToLower(err.Error()), []string{"database not found"}) {
+		if utils.ContainsAny(strings.ToLower(err.Error()), []string{"database not found"}) {
 			return false, nil
 		}
 		return false, fmt.Errorf("can't get database info: %s", err)
@@ -534,7 +395,7 @@ func CreateOrUpdateDatabase(ctx context.Context, adminClient *database.DatabaseA
 // Spanner instance to use, generates a new Spanner DB name,
 // and call into the Spanner admin interface to create the new DB.
 func CreateDatabase(ctx context.Context, adminClient *database.DatabaseAdminClient, dbURI string, conv *internal.Conv, out *os.File) error {
-	project, instance, dbName := parseDbURI(dbURI)
+	project, instance, dbName := utils.ParseDbURI(dbURI)
 	fmt.Fprintf(out, "Creating new database %s in instance %s with default permissions ... \n", dbName, instance)
 	// The schema we send to Spanner excludes comments (since Cloud
 	// Spanner DDL doesn't accept them), and protects table and col names
@@ -557,10 +418,10 @@ func CreateDatabase(ctx context.Context, adminClient *database.DatabaseAdminClie
 
 	op, err := adminClient.CreateDatabase(ctx, req)
 	if err != nil {
-		return fmt.Errorf("can't build CreateDatabaseRequest: %w", AnalyzeError(err, dbURI))
+		return fmt.Errorf("can't build CreateDatabaseRequest: %w", utils.AnalyzeError(err, dbURI))
 	}
 	if _, err := op.Wait(ctx); err != nil {
-		return fmt.Errorf("createDatabase call failed: %w", AnalyzeError(err, dbURI))
+		return fmt.Errorf("createDatabase call failed: %w", utils.AnalyzeError(err, dbURI))
 	}
 	fmt.Fprintf(out, "Created database successfully.\n")
 
@@ -585,46 +446,13 @@ func UpdateDatabase(ctx context.Context, adminClient *database.DatabaseAdminClie
 	}
 	op, err := adminClient.UpdateDatabaseDdl(ctx, req)
 	if err != nil {
-		return fmt.Errorf("can't build UpdateDatabaseDdlRequest: %w", AnalyzeError(err, dbURI))
+		return fmt.Errorf("can't build UpdateDatabaseDdlRequest: %w", utils.AnalyzeError(err, dbURI))
 	}
 	if err := op.Wait(ctx); err != nil {
-		return fmt.Errorf("UpdateDatabaseDdl call failed: %w", AnalyzeError(err, dbURI))
+		return fmt.Errorf("UpdateDatabaseDdl call failed: %w", utils.AnalyzeError(err, dbURI))
 	}
 	fmt.Fprintf(out, "Updated schema successfully.\n")
 	return nil
-}
-
-// parseURI parses an unknown URI string that could be a database, instance or project URI.
-func parseURI(URI string) (project, instance, dbName string) {
-	project, instance, dbName = "", "", ""
-	if strings.Contains(URI, "databases") {
-		project, instance, dbName = parseDbURI(URI)
-	} else if strings.Contains(URI, "instances") {
-		project, instance = parseInstanceURI(URI)
-	} else if strings.Contains(URI, "projects") {
-		project = parseProjectURI(URI)
-	}
-	return
-}
-
-func parseDbURI(dbURI string) (project, instance, dbName string) {
-	split := strings.Split(dbURI, "/databases/")
-	project, instance = parseInstanceURI(split[0])
-	dbName = split[1]
-	return
-}
-
-func parseInstanceURI(instanceURI string) (project, instance string) {
-	split := strings.Split(instanceURI, "/instances/")
-	project = parseProjectURI(split[0])
-	instance = split[1]
-	return
-}
-
-func parseProjectURI(projectURI string) (project string) {
-	split := strings.Split(projectURI, "/")
-	project = split[1]
-	return
 }
 
 // UpdateDDLForeignKeys updates the Spanner database with foreign key
@@ -696,72 +524,6 @@ Recommended value is between 20-30.`)
 	}
 	p.Done()
 	return nil
-}
-
-// GetProject returns the cloud project we should use for accessing Spanner.
-// Use environment variable GCLOUD_PROJECT if it is set.
-// Otherwise, use the default project returned from gcloud.
-func GetProject() (string, error) {
-	project := os.Getenv("GCLOUD_PROJECT")
-	if project != "" {
-		return project, nil
-	}
-	cmd := exec.Command("gcloud", "config", "list", "--format", "value(core.project)")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return "", fmt.Errorf("call to gcloud to get project failed: %w", err)
-	}
-	project = strings.TrimSpace(string(out))
-	return project, nil
-}
-
-// GetInstance returns the Spanner instance we should use for creating DBs.
-// If the user specified instance (via flag 'instance') then use that.
-// Otherwise try to deduce the instance using gcloud.
-func GetInstance(ctx context.Context, project string, out *os.File) (string, error) {
-	l, err := getInstances(ctx, project)
-	if err != nil {
-		return "", err
-	}
-	if len(l) == 0 {
-		fmt.Fprintf(out, "Could not find any Spanner instances for project %s\n", project)
-		return "", fmt.Errorf("no Spanner instances for %s", project)
-	}
-
-	// Note: we could ask for user input to select/confirm which Spanner
-	// instance to use, but that interacts poorly with piping pg_dump/mysqldump data
-	// to the tool via stdin.
-	if len(l) == 1 {
-		fmt.Fprintf(out, "Using only available Spanner instance: %s\n", l[0])
-		return l[0], nil
-	}
-	fmt.Fprintf(out, "Available Spanner instances:\n")
-	for i, x := range l {
-		fmt.Fprintf(out, " %d) %s\n", i+1, x)
-	}
-	fmt.Fprintf(out, "Please pick one of the available instances and set the flag '--instance'\n\n")
-	return "", fmt.Errorf("auto-selection of instance failed: project %s has more than one Spanner instance. "+
-		"Please use the flag '--instance' to select an instance", project)
-}
-
-func getInstances(ctx context.Context, project string) ([]string, error) {
-	instanceClient, err := instance.NewInstanceAdminClient(ctx)
-	if err != nil {
-		return nil, AnalyzeError(err, fmt.Sprintf("projects/%s", project))
-	}
-	it := instanceClient.ListInstances(ctx, &instancepb.ListInstancesRequest{Parent: fmt.Sprintf("projects/%s", project)})
-	var l []string
-	for {
-		resp, err := it.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return nil, AnalyzeError(err, fmt.Sprintf("projects/%s", project))
-		}
-		l = append(l, strings.TrimPrefix(resp.Name, fmt.Sprintf("projects/%s/instances/", project)))
-	}
-	return l, nil
 }
 
 // WriteSchemaFile writes DDL statements in a file. It includes CREATE TABLE
@@ -878,7 +640,7 @@ func ReadSessionFile(conv *internal.Conv, sessionJSON string) error {
 // to file 'name'.
 func WriteBadData(bw *spanner.BatchWriter, conv *internal.Conv, banner, name string, out *os.File) {
 	badConversions := conv.BadRows()
-	badWrites := sum(bw.DroppedRowsByTable())
+	badWrites := utils.SumMapValues(bw.DroppedRowsByTable())
 	if badConversions == 0 && badWrites == 0 {
 		os.Remove(name) // Cleanup bad-data file from previous run.
 		return
@@ -921,167 +683,6 @@ func WriteBadData(bw *spanner.BatchWriter, conv *internal.Conv, banner, name str
 		}
 	}
 	fmt.Fprintf(out, "See file '%s' for details of bad rows\n", name)
-}
-
-// GetDatabaseName generates database name with driver_date prefix.
-func GetDatabaseName(driver string, now time.Time) (string, error) {
-	return generateName(fmt.Sprintf("%s_%s", driver, now.Format("2006-01-02")))
-}
-
-func GetPassword() string {
-	fmt.Print("Enter Password: ")
-	bytePassword, err := terminal.ReadPassword(int(syscall.Stdin))
-	if err != nil {
-		fmt.Println("\nCoudln't read password")
-		return ""
-	}
-	fmt.Printf("\n")
-	return strings.TrimSpace(string(bytePassword))
-}
-
-// AnalyzeError inspects an error returned from Cloud Spanner and adds information
-// about potential root causes e.g. authentication issues.
-func AnalyzeError(err error, URI string) error {
-	project, instance, _ := parseURI(URI)
-	e := strings.ToLower(err.Error())
-	if containsAny(e, []string{"unauthenticated", "cannot fetch token", "default credentials"}) {
-		return fmt.Errorf("%w."+`
-Possible cause: credentials are mis-configured. Do you need to run
-
-  gcloud auth application-default login
-
-or configure environment variable GOOGLE_APPLICATION_CREDENTIALS.
-See https://cloud.google.com/docs/authentication/getting-started`, err)
-	}
-	if containsAny(e, []string{"instance not found"}) && instance != "" {
-		return fmt.Errorf("%w.\n"+`
-Possible cause: Spanner instance specified via instance option does not exist.
-Please check that '%s' is correct and that it is a valid Spanner
-instance for project %s`, err, instance, project)
-	}
-	return err
-}
-
-// PrintPermissionsWarning prints permission warning.
-func PrintPermissionsWarning(driver string, out *os.File) {
-	fmt.Fprintf(out,
-		`
-WARNING: Please check that permissions for this Spanner instance are
-appropriate. Spanner manages access control at the database level, and the
-database created by HarbourBridge will inherit default permissions from this
-instance. All data written to Spanner will be visible to anyone who can
-access the created database. Note that `+driver+` table-level and row-level
-ACLs are dropped during conversion since they are not supported by Spanner.
-
-`)
-}
-
-func printSeekError(driver string, err error, out *os.File) {
-	fmt.Fprintf(out, "\nCan't get seekable input file: %v\n", err)
-	fmt.Fprintf(out, "Likely cause: not enough space in %s.\n", os.TempDir())
-	fmt.Fprintf(out, "Try writing "+driver+" output to a file first i.e.\n")
-	fmt.Fprintf(out, " "+driver+" > tmpfile\n")
-	fmt.Fprintf(out, "  harbourbridge < tmpfile\n")
-}
-
-func containsAny(s string, l []string) bool {
-	for _, a := range l {
-		if strings.Contains(s, a) {
-			return true
-		}
-	}
-	return false
-}
-
-func generateName(prefix string) (string, error) {
-	b := make([]byte, 4)
-	_, err := rand.Read(b)
-	if err != nil {
-		return "", fmt.Errorf("error generating name: %w", err)
-
-	}
-	return fmt.Sprintf("%s_%x-%x", prefix, b[0:2], b[2:4]), nil
-}
-
-// NewSpannerClient returns a new Spanner client.
-// It respects SPANNER_API_ENDPOINT.
-func NewSpannerClient(ctx context.Context, db string) (*sp.Client, error) {
-	if endpoint := os.Getenv("SPANNER_API_ENDPOINT"); endpoint != "" {
-		return sp.NewClient(ctx, db, option.WithEndpoint(endpoint))
-	}
-	return sp.NewClient(ctx, db)
-}
-
-// GetClient returns a new Spanner client.  It uses the background context.
-func GetClient(ctx context.Context, db string) (*sp.Client, error) {
-	return NewSpannerClient(ctx, db)
-}
-
-// NewDatabaseAdminClient returns a new db-admin client.
-// It respects SPANNER_API_ENDPOINT.
-func NewDatabaseAdminClient(ctx context.Context) (*database.DatabaseAdminClient, error) {
-	if endpoint := os.Getenv("SPANNER_API_ENDPOINT"); endpoint != "" {
-		return database.NewDatabaseAdminClient(ctx, option.WithEndpoint(endpoint))
-	}
-	return database.NewDatabaseAdminClient(ctx)
-}
-
-// NewInstanceAdminClient returns a new instance-admin client.
-// It respects SPANNER_API_ENDPOINT.
-func NewInstanceAdminClient(ctx context.Context) (*instance.InstanceAdminClient, error) {
-	if endpoint := os.Getenv("SPANNER_API_ENDPOINT"); endpoint != "" {
-		return instance.NewInstanceAdminClient(ctx, option.WithEndpoint(endpoint))
-	}
-	return instance.NewInstanceAdminClient(ctx)
-}
-
-func getSize(f *os.File) (int64, error) {
-	info, err := f.Stat()
-	if err != nil {
-		return 0, fmt.Errorf("can't stat file: %w", err)
-	}
-	return info.Size(), nil
-}
-
-// SetupLogFile configures the file used for logs.
-// By default we just drop logs on the floor. To enable them (e.g. to debug
-// Cloud Spanner client library issues), set logfile to a non-empty filename.
-// Note: this tool itself doesn't generate logs, but some of the libraries it
-// uses do. If we don't set the log file, we see a number of unhelpful and
-// unactionable logs spamming stdout, which is annoying and confusing.
-func SetupLogFile() (*os.File, error) {
-	// To enable debug logs, set logfile to a non-empty filename.
-	logfile := ""
-	if logfile == "" {
-		log.SetOutput(ioutil.Discard)
-		return nil, nil
-	}
-	f, err := os.Create(logfile)
-	if err != nil {
-		return nil, err
-	}
-	log.SetOutput(f)
-	return f, nil
-}
-
-// Close closes file.
-func Close(f *os.File) {
-	if f != nil {
-		f.Close()
-	}
-}
-
-func sum(m map[string]int64) int64 {
-	n := int64(0)
-	for _, c := range m {
-		n += c
-	}
-	return n
-}
-
-// GetBanner prints banner message after command line process is finished.
-func GetBanner(now time.Time, db string) string {
-	return fmt.Sprintf("Generated at %s for db %s\n\n", now.Format("2006-01-02 15:04:05"), db)
 }
 
 // ProcessDump invokes process dump function from a sql package based on driver selected.
