+++ profiles/target_profile.go
@@ -1,4 +1,4 @@
-package cmd
+package profiles
 
 import (
 	"fmt"
